Eliminated a compiler warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@2189 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csengine/terrain.cpp
@@ -176,7 +176,7 @@ void csTerrain::Draw (csRenderView& rview, bool use_z_buf)
 			p = bt->tri(tvp)->pos();
             if (p->v[2] < d)
                 d = p->v[2];
-            mesh->qz()->ddgSplayTree::insert (bt->index(),tvc,d);
+            mesh->qz()->ddgSplayTree::insert (bt->index(),tvc,(unsigned int)d);
 			mesh->qsi ()->next ();
 		}
 		qri = mesh->qzi();
